Fixed parallel performance by not doing simpF in parallel.
Note that simpF is not the slow part. Also it helps cancellation between
terms. If simpF is done in parallel, cancellation between terms on
different kernels is no longer possible.

One can do two passes and first do simpF on parallel. But there should
be no point to do it considering it is fast.